make_bullet
